PDFBOX-5889: support long values for COSInteger
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/branches/2.0@1921498 13f79535-47bb-0310-9956-ffa450edef68

+++ debugger/src/main/java/org/apache/pdfbox/debugger/ui/PDFTreeCellRenderer.java
@@ -133,7 +133,7 @@ public class PDFTreeCellRenderer extends DefaultTreeCellRenderer
         }
         else if (nodeValue instanceof COSInteger)
         {
-            result = Integer.toString(((COSInteger) nodeValue).intValue());
+            result = Long.toString(((COSInteger) nodeValue).longValue());
         }
         else if (nodeValue instanceof COSString)
         {
